log-service/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
pydantic==1.8.2
kafka-python==2.0.2 